Added .skel File Saving

+++ FileSaver.py
@@ -6,15 +6,10 @@ import json
 import traceback as tb
 
 
-def saveToFile(dir, fileName, frameMan):
+def saveToJson(dir, fileName, frameMan):
     fileName = os.path.splitext(fileName)[0] + ".json"
 
     if not os.path.exists(dir):
-        print("Directory {} does not exist".format(dir))
-        raise FileNotFoundError("Directory {} does not exist".format(dir))
-
-    if not os.path.exists(dir):
-        print("Directory {} does not exist".format(dir))
         raise FileNotFoundError("Directory {} does not exist".format(dir))
 
     if not isinstance(frameMan, (FrameManager)):
@@ -23,11 +18,34 @@ def saveToFile(dir, fileName, frameMan):
     filePath = os.path.join(dir, fileName)
     tempPath = filePath + ".temp"
     try:
-        # with open(tempPath, "wb") as outFile:
-        #     pickle.dump(frameMan.getPickles(), outFile)
-
         with open(tempPath, "w") as outFile:
-            json.dump(frameMan.makeDict(), outFile)
+            json.dump(frameMan.makeDict(), outFile, indent=2)
+
+        if os.path.exists(filePath):
+            os.remove(filePath)
+        os.rename(tempPath, filePath)
+
+    except Exception as e:
+        if os.path.exists(tempPath):
+            os.remove(tempPath)
+        print(str(e))
+        print(tb.format_exc())
+
+def saveToBinary(dir, fileName, frameMan):
+    fileName = os.path.splitext(fileName)[0] + ".skel"
+
+    if not os.path.exists(dir):
+        raise FileNotFoundError("Directory {} does not exist".format(dir))
+
+    if not isinstance(frameMan, (FrameManager)):
+        raise ValueError("Please Provide frameManager To Save")
+
+    filePath = os.path.join(dir, fileName)
+    tempPath = filePath + ".temp"
+    try:
+
+        with open(tempPath, "wb") as outFile:
+            outFile.write(frameMan.asBytes())
 
         if os.path.exists(filePath):
             os.remove(filePath)
